services/ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pytesseract==0.3.10
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.2
